FastAPI-server/requirements.txt
fastapi==0.111.0
joblib==1.3.2
nltk==3.8.1
pandas==2.2.2
pydantic==2.8.0
scikit_learn==1.4.1.post1
uvicorn==0.30.1
python-dotenv==1.0.1
httpx==0.27.0
bson==0.5.10
pymongo==4.8.0
passlib==1.7.4
PyJWT==2.8.0
google-generativeai==0.7.2
replicate==0.30.1